\name{do}
\alias{do}
\alias{magic.wand}
\title{
Apply a function f to each of the groups of a data set, or set up a method to do just that.
}
\description{
\code{do} takes a pipe and extends it so that when the pipe is executed the function f is applied to each of the groups in the data set generated by .data. The return values of such application become the output of the extended pipe.  \code{magic.wand} creates a method for class pipe for a function named like \code{f} which does exactly what do does. 
}
\usage{
do(.data, .f, \dots)
magic.wand(f)
}
\arguments{
  \item{.data}{A pipe}
  \item{f,.f}{A function of a data frame returning a data frame (attempts are made to coerce a non data frame return value to a data frame, but your milage may vary)}
  \item{\dots}{Additional arguments to f}
}
\details{
If you have a function tha acts on and returns data frames and want to make it into a generic function by the same name that will do whatever it used to do when called on data frames but instead call \code{do(.data, f, ...)} when called on a pipe, just call \code{magic.wand} and it will happen. For instance, \code{melt} is a generic function from reshape2 that didn't have a method for pipes. By calling \code{magic.wand} on it, a method as been added for pipes. 
}
\value{
A pipe extended with the application of f to each group.
}
\examples{
}
