\name{plyrmr-package}
\alias{plyrmr-package}
\alias{plyrmr}
\docType{package}
\title{
What the package does (short line)
}
\description{
Easy big data manipulation a la plyr with R and Hadoop 
}
\details{
\tabular{ll}{
Package: \tab plyrmr\cr
Type: \tab Package\cr
Version: \tab 0.0.0\cr
Date: \tab any day\cr
License: \tab Apache 2.0\cr
}
}
\author{rhadoop@revolutionanalytics.com}
