\name{sample}
\alias{sample}

\title{
Function to sample from big data sets and data frames
}
\description{
Albeit methods for a generic function whose default method is the well known base::sample, the methods for data frames and pipes have a different behavior, but they are still meant to perform sampling,
}
\usage{
sample(...)
\method{sample}{pipe}(x, method = c("any", "Bernoulli"))
\method{sample}{data.frame}(x, method = c("any", "Bernoulli"))
}
\arguments{
  \item{method}
  \item{\dots}{}
}
\details{
}
\value{
}
\examples{
}
