\name{where}
\alias{where}
\alias{select}
\title{
Functions to perform sql-like operations on data frames and pipes.
}
\description{

}
\usage{
where(.data, ...)
\method{where}{data.frame}(.data, ...)
\method{where}{pipe}(.data, ...)
select(.data, ...)
\method{select}{data.frame}(.data, ...)
\method{select}{pipe}(.data, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{.data}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\examples{
}
