\name{plyrmr-package}
\alias{plyrmr-package}
\alias{plyrmr}
\docType{package}
\title{
Easy big data manipulation plyr-style with R and Hadoop }
\description{
}
\details{
\tabular{ll}{
Package: \tab plyrmr\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab any day\cr
License: \tab Apache 2.0\cr
}
}
\author{rhadoop@revolutionanalytics.com}
