\name{dcast}
\alias{dcast}
\alias{melt}
\title{
Function to convert big data sets from tall to wide and back.
}
\description{
Modeled after, and actually implemented on top of, equally named functions in the legendary packages \code{reshape2}, they implement the same functionality for large data sets.
}
\usage{
dcast(data, \dots)
\method{dcast}{pipe, formula, \dots}
melt(data, \dots)
\method{melt}{pipe, \dots}
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{}
  \item{formula}{}
  \item{\dots}{}
}

\details{
}
\value{A pipe with the desired format
}
\examples{
}
