\name{where}
\alias{where}
\alias{select}
\title{
Functions to perform sql-like operations on data frames and pipes.
}
\description{

}
\usage{
where(.data, \dots)
\method{where}{data.frame}(.data, \dots)
\method{where}{pipe}(.data, \dots)
select(.data, \dots)
\method{select}{data.frame}(.data, \dots)
\method{select}{pipe}(.data, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{.data}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\examples{
}
